refactor: keep image ui with its feature
